# src/tide-tabular/tests/fixtures/tabular.openapi.json
{
  "components": {
    "schemas": {
...
      "ColumnDTO": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "numeric",
              "categorical",
              "text"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "ColumnDTO",
        "type": "object"
      },
...
      "ExecutionDTO": {
        "additionalProperties": false,
        "properties": {
          "deadline_ms": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Ms"
          },
          "policy": {
            "const": "no_replay",
            "default": "no_replay",
            "title": "Policy",
            "type": "string"
          }
        },
        "title": "ExecutionDTO",
        "type": "object"
      },
...
      "FitRequestV3DTO": {
        "additionalProperties": false,
        "properties": {
          "backendKey": {
            "enum": [
              "fake",
              "priorlabs",
              "aicore"
            ],
            "title": "Backendkey",
            "type": "string"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnDTO"
            },
            "minItems": 1,
            "title": "Columns",
            "type": "array"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionDTO"
          },
          "fitOptions": {
            "$ref": "#/components/schemas/FitOptionsDTO"
          },
          "modelKey": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Modelkey",
            "type": "string"
          },
          "modelOptions": {
            "$ref": "#/components/schemas/ModelOptionsDTO"
          },
          "task": {
            "enum": [
              "classification",
              "regression"
            ],
            "title": "Task",
            "type": "string"
          },
          "x_train": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "minItems": 1,
            "title": "X Train",
            "type": "array"
          },
          "y_train": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "minItems": 1,
            "title": "Y Train",
            "type": "array"
          }
        },
        "required": [
          "task",
          "backendKey",
          "modelKey",
          "columns",
          "x_train",
          "y_train"
        ],
        "title": "FitRequestV3DTO",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelReferenceDTO": {
        "additionalProperties": false,
        "properties": {
          "backendKey": {
            "title": "Backendkey",
            "type": "string"
          },
          "configuration_identity": {
            "title": "Configuration Identity",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "modelKey": {
            "title": "Modelkey",
            "type": "string"
          },
          "token": {
            "maxLength": 131072,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          },
          "version": {
            "const": 1,
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "backendKey",
          "modelKey",
          "configuration_identity",
          "token"
        ],
        "title": "ModelReferenceDTO",
        "type": "object"
      },
      "OutputDTO": {
        "additionalProperties": false,
        "properties": {
          "levels": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Levels"
          },
          "type": {
            "enum": [
              "probas",
              "point",
              "quantiles"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "OutputDTO",
        "type": "object"
      },
...
      "PredictRequestDTO": {
        "additionalProperties": false,
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnDTO"
            },
            "minItems": 1,
            "title": "Columns",
            "type": "array"
          },
          "keys": {
            "items": {
              "type": "string"
            },
            "title": "Keys",
            "type": "array"
          },
          "mode": {
            "default": "predict",
            "enum": [
              "predict",
              "dry_run"
            ],
            "title": "Mode",
            "type": "string"
          },
          "output": {
            "$ref": "#/components/schemas/OutputDTO"
          },
          "task": {
            "enum": [
              "classification",
              "regression"
            ],
            "title": "Task",
            "type": "string"
          },
          "x_test": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "X Test",
            "type": "array"
          },
          "x_train": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "X Train",
            "type": "array"
          },
          "y_train": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Y Train",
            "type": "array"
          }
        },
        "required": [
          "task",
          "columns",
          "x_train",
          "y_train",
          "keys",
          "x_test",
          "output"
        ],
        "title": "PredictRequestDTO",
        "type": "object"
      },
...
      "ReleaseRequestDTO": {
        "additionalProperties": false,
        "properties": {
          "modelRef": {
            "$ref": "#/components/schemas/ModelReferenceDTO"
          }
        },
        "required": [
          "modelRef"
        ],
        "title": "ReleaseRequestDTO",
        "type": "object"
      },
      "ReleaseResultDTO": {
        "properties": {
          "released": {
            "const": false,
            "default": false,
            "title": "Released",
            "type": "boolean"
          },
          "status": {
            "const": "unsupported",
            "default": "unsupported",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "ReleaseResultDTO",
        "type": "object"
      },
...
      "TabularResultDTO": {
        "properties": {
          "classes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classes"
          },
          "dropped_columns": {
            "items": {
              "type": "string"
            },
            "title": "Dropped Columns",
            "type": "array"
          },
          "elapsed_ms": {
            "title": "Elapsed Ms",
            "type": "number"
          },
          "fallback": {
            "anyOf": [
              {
                "enum": [
                  "context_distribution",
                  "context_quantiles"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback"
          },
          "levels": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Levels"
          },
          "output_type": {
            "enum": [
              "probas",
              "point",
              "quantiles"
            ],
            "title": "Output Type",
            "type": "string"
          },
          "placeholder": {
            "title": "Placeholder",
            "type": "boolean"
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/PredictionDTO"
            },
            "title": "Predictions",
            "type": "array"
          },
          "task": {
            "enum": [
              "classification",
              "regression"
            ],
            "title": "Task",
            "type": "string"
          },
          "train_rows": {
            "title": "Train Rows",
            "type": "integer"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageDTO"
          }
        },
        "required": [
          "task",
          "output_type",
          "predictions",
          "dropped_columns",
          "placeholder",
          "usage",
          "train_rows",
          "elapsed_ms"
        ],
        "title": "TabularResultDTO",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "tabular",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tabular": {
      "post": {
        "description": "`X-Deadline-Ms` bounds worker waiting, relative (no clock sync needed).\nProvider work may outlive it; parsing and preprocessing are not a hard SLA.",
        "operationId": "predict_v1_tabular_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-deadline-ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Deadline-Ms"
            }
          },
          {
            "in": "header",
            "name": "x-backend-identity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Backend-Identity"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TabularResultDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict"
      }
    },
...
    "/v3/models/fit": {
      "post": {
        "operationId": "fit_model_v3_v3_models_fit_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-backend-identity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Backend-Identity"
            }
          },
          {
            "in": "header",
            "name": "x-deadline-ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Deadline-Ms"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitRequestV3DTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelReferenceDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fit Model V3"
      }
    },
    "/v3/models/release": {
      "post": {
        "operationId": "release_model_v3_v3_models_release_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResultDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Model V3"
      }
    },
    "/v3/tabular": {
      "post": {
        "operationId": "predict_v3_v3_tabular_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-deadline-ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Deadline-Ms"
            }
          },
          {
            "in": "header",
            "name": "x-backend-identity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Backend-Identity"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequestV3DTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TabularResultV3DTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict V3"
      }
    }
  }
}